[tests] cleaning up unit tests in preparation of adding alert processor unit tests
