+++ test/scripts/unit_tests.sh
@@ -1,2 +1,2 @@
 #! /bin/bash
-nosetests test/unit --with-coverage --cover-package=stream_alert --cover-package=stream_alert_cli --cover-min-percentage=80
+nosetests test/unit --with-coverage --cover-package=stream_alert.rule_processor --cover-package=stream_alert_cli --cover-min-percentage=80
