+++ test/unit/stream_alert_rule_processor/test_gzip_json_parser.py
@@ -1,38 +1,44 @@
-from stream_alert.rule_processor.config import load_config
-from stream_alert.rule_processor.parsers import get_parser
+'''
+Copyright 2017-present, Airbnb Inc.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+   http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+'''
 import zlib
 
-from nose.tools import (
-    assert_equal,
-    assert_not_equal,
-    nottest,
-    assert_raises,
-    raises
-)
+from nose.tools import assert_equal, assert_not_equal
+
+from stream_alert.rule_processor.config import load_config
+from stream_alert.rule_processor.parsers import get_parser
+
 
 class TestGzipJsonParser(object):
+    """Test class for GZIP JSON parser"""
     @classmethod
     def setup_class(cls):
-        """setup_class() before any methods in this class"""
-        pass
+        """Setup the class before any methods"""
+        # load config
+        cls.config = load_config('test/unit/conf')
+        # load JSON parser class
+        cls.parser_class = get_parser('gzip-json')
 
     @classmethod
     def teardown_class(cls):
-        """teardown_class() after any methods in this class"""
-        pass
-
-    def setup(self):
-        """Setup before each method"""
-        # load config
-        self.config = load_config('test/unit/conf')
-        # load JSON parser class
-        self.parser_class = get_parser('gzip-json')
-
-    def teardown(self):
-        """Teardown after each method"""
-        pass
+        """Teardown the class after all methods"""
+        cls.config = None
+        cls.parser_class = None
 
     def parser_helper(self, **kwargs):
+        """Helper to return the parser result"""
         data = kwargs['data']
         schema = kwargs['schema']
         options = kwargs['options']
@@ -46,7 +52,7 @@ class TestGzipJsonParser(object):
         schema = self.config['logs']['test_cloudwatch']['schema']
         options = self.config['logs']['test_cloudwatch']['configuration']
 
-        with open('test/unit/fixtures/cloudwatch.json','r') as fixture_file:
+        with open('test/unit/fixtures/cloudwatch.json', 'r') as fixture_file:
             data = fixture_file.readlines()
         data_record = zlib.compress(data[0].strip())
 
@@ -55,14 +61,15 @@ class TestGzipJsonParser(object):
                                            options=options)
 
         assert_not_equal(parsed_result, False)
-        assert_equal(80,len(parsed_result))
+        assert_equal(80, len(parsed_result))
 
         expected_keys = (u'protocol', u'source', u'destination', u'srcport',
                          u'destport', u'eni', u'action', u'packets', u'bytes',
                          u'windowstart', u'windowend', u'version', u'account',
-                         u'flowlogstatus',u'streamalert:envelope_keys')
+                         u'flowlogstatus', u'streamalert:envelope_keys')
         expected_envelope_keys = (u'logGroup', u'logStream', u'owner')
 
         for result in parsed_result:
             assert_equal(sorted(expected_keys), sorted(result.keys()))
-            assert_equal(sorted(expected_envelope_keys),sorted(result['streamalert:envelope_keys'].keys()))
+            assert_equal(sorted(expected_envelope_keys),
+                         sorted(result['streamalert:envelope_keys'].keys()))
